HDDS-11768. Extract SCM failover proxy provider logic
